Move image command handler and related services into proper namespace
