+++ src/Adapter/Product/Image/Uploader/ProductImageUploader.php
@@ -26,13 +26,14 @@
 
 declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Adapter\Image\Uploader;
+namespace PrestaShop\PrestaShop\Adapter\Product\Image\Uploader;
 
 use ErrorException;
 use Image;
 use PrestaShop\PrestaShop\Adapter\Image\ImageGenerator;
-use PrestaShop\PrestaShop\Adapter\Product\ProductImagePathFactory;
-use PrestaShop\PrestaShop\Adapter\Product\Repository\ProductImageRepository;
+use PrestaShop\PrestaShop\Adapter\Image\Uploader\AbstractImageUploader;
+use PrestaShop\PrestaShop\Adapter\Product\Image\ProductImagePathFactory;
+use PrestaShop\PrestaShop\Adapter\Product\Image\Repository\ProductImageRepository;
 use PrestaShop\PrestaShop\Core\Configuration\UploadSizeConfigurationInterface;
 use PrestaShop\PrestaShop\Core\Hook\HookDispatcherInterface;
 use PrestaShop\PrestaShop\Core\Image\Exception\CannotUnlinkImageException;
@@ -76,13 +77,13 @@ class ProductImageUploader extends AbstractImageUploader
     private $isLegacyImageMode;
 
     /**
-     * @var ProductImageRepository
+     * @var \PrestaShop\PrestaShop\Adapter\Product\Image\Repository\ProductImageRepository
      */
     private $productImageRepository;
 
     /**
      * @param UploadSizeConfigurationInterface $uploadSizeConfiguration
-     * @param ProductImagePathFactory $productImagePathFactory
+     * @param \PrestaShop\PrestaShop\Adapter\Product\Image\ProductImagePathFactory $productImagePathFactory
      * @param int $contextShopId
      * @param ImageGenerator $imageGenerator
      * @param HookDispatcherInterface $hookDispatcher
