+++ src/PrestaShopBundle/Resources/config/services/adapter/image.yml
@@ -36,14 +36,3 @@ services:
 
     prestashop.adapter.image.image_validator:
         class: 'PrestaShop\PrestaShop\Adapter\Image\ImageValidator'
-
-    prestashop.adapter.image.uploader.product_image_uploader:
-        class: 'PrestaShop\PrestaShop\Adapter\Image\Uploader\ProductImageUploader'
-        arguments:
-            - '@prestashop.core.configuration.upload_size_configuration'
-            - '@prestashop.adapter.product.product_image_path_factory'
-            - "@=service('prestashop.adapter.legacy.context').getContext().shop.id"
-            - "@prestashop.adapter.image.image_generator"
-            - "@prestashop.core.hook.dispatcher"
-            - "@=service('prestashop.adapter.legacy.configuration').getBoolean('PS_LEGACY_IMAGES')"
-            - '@prestashop.adapter.product.repository.product_image_repository'
