+++ src/PrestaShopBundle/Resources/config/services/adapter/product/_image.yml
@@ -1,0 +1,38 @@
+services:
+    _defaults:
+        public: true
+
+    prestashop.adapter.product.image.command_handler.add_product_image_handler:
+        class: PrestaShop\PrestaShop\Adapter\Product\Image\CommandHandler\AddProductImageHandler
+        arguments:
+            - '@prestashop.adapter.product.image.uploader.product_image_uploader'
+            - '@prestashop.adapter.product.image.repository.product_image_repository'
+            - "@prestashop.adapter.image.image_validator"
+            - "@=service('prestashop.adapter.shop.context').getContextListShopID()"
+        tags:
+            - name: tactician.handler
+              command: PrestaShop\PrestaShop\Core\Domain\Product\Image\Command\AddProductImageCommand
+
+    prestashop.adapter.product.image.repository.product_image_repository:
+        class: PrestaShop\PrestaShop\Adapter\Product\Image\Repository\ProductImageRepository
+        arguments:
+            - '@doctrine.dbal.default_connection'
+            - '%database_prefix%'
+
+    prestashop.adapter.product.image.product_image_path_factory:
+        class: PrestaShop\PrestaShop\Adapter\Product\Image\ProductImagePathFactory
+        arguments:
+            - "@=service('prestashop.adapter.legacy.configuration').getBoolean('PS_LEGACY_IMAGES')"
+            - !php/const _PS_TMP_IMG_DIR_
+
+    prestashop.adapter.product.image.uploader.product_image_uploader:
+        class: 'PrestaShop\PrestaShop\Adapter\Product\Image\Uploader\ProductImageUploader'
+        arguments:
+            - '@prestashop.core.configuration.upload_size_configuration'
+            - '@prestashop.adapter.product.image.product_image_path_factory'
+            - "@=service('prestashop.adapter.legacy.context').getContext().shop.id"
+            - "@prestashop.adapter.image.image_generator"
+            - "@prestashop.core.hook.dispatcher"
+            - "@=service('prestashop.adapter.legacy.configuration').getBoolean('PS_LEGACY_IMAGES')"
+            - '@prestashop.adapter.product.image.repository.product_image_repository'
+
